% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_SHGInterface}
\alias{Rcpp_SHGInterface}
\title{Rcpp SHG Interface Class}
\description{
This module provides an Rcpp interface to the Smoking History Generator (SHG) application.
}
\details{
Rcpp SHG Interface Class
}
