\name{RcppSmokingHistoryGenerator-package}
\alias{RcppSmokingHistoryGenerator-package}
\alias{RcppSmokingHistoryGenerator}
\docType{package}
\title{
\packageTitle{RcppSmokingHistoryGenerator}
}
\description{
\packageDescription{RcppSmokingHistoryGenerator}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{RcppSmokingHistoryGenerator}

Maintainer: \packageMaintainer{RcppSmokingHistoryGenerator}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = RcppSmokingHistoryGenerator)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{RcppSmokingHistoryGenerator}}
%\section{Documentation Index}{\packageIndices{RcppSmokingHistoryGenerator}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
