% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RunWebVersion}
\alias{RunWebVersion}
\title{RunWebVersion}
\usage{
RunWebVersion()
}
\value{
bool
}
\description{
Executes a simluation based on parameters from a single input file
}
\examples{
RunWebVersion()
}
