% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_params.R
\name{shg_params_cache_dir}
\alias{shg_params_cache_dir}
\title{Return the directory where downloaded parameter sets are cached}
\usage{
shg_params_cache_dir()
}
\description{
Return the directory where downloaded parameter sets are cached
}
