% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LegacyRunWebVersion}
\alias{LegacyRunWebVersion}
\title{LegacyRunWebVersion method}
\arguments{
\item{input_file_name}{The name of the configuration file (see ./inst/inputs/ for 2 examples)}
}
\description{
This method offers a way to configure and run a simulation from an input configuration file. Rather than return a R DataFrame, it produces results in an output file. It works in the same as calling the CLI version of the Smoking History Generator with a single input file parameter.
}
\examples{
# Warning: This way of running a simulation ignores the Rcpp interface properties and relies soley 
# on parameters set in the input configuration file. See main.cpp's RunWebVersion for more detail.
library(RcppSmokingHistoryGenerator)

shg <- new(SHGInterface)
shg$LegacyRunWebVersion("/inst/inputs/test_input_example_MersenneTwister.txt")
}
