\name{Rcpp_SHGInterface-class}
\docType{class}
\alias{Rcpp_SHGInterface-class}

\title{Class \code{"Rcpp_SHGInterface"}}
\description{
This is a description.
}
\section{Extends}{
Class \code{"\link[Rcpp:C++Object-class]{C++Object}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.

}
\references{
These are my references
}
\author{
I am the author
}
\note{
Further Notes
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
See also Sample
}
\examples{
showClass("Rcpp_SHGInterface")
}
\keyword{classes}
