\name{Rcpp_SHGInterface-class}
\docType{class}
\alias{Rcpp_SHGInterface-class}

\title{Class \code{"Rcpp_SHGInterface"}}
\description{
This is a description.
}
\section{Extends}{
Class \code{"C++Object"} (Rcpp base class; see \pkg{Rcpp}), directly.

All reference classes extend the S4 virtual class \code{envRefClass}; see \pkg{methods}.

}
\references{
These are my references
}
\author{
I am the author
}
\note{
Further Notes
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
See also Sample
}
\examples{
showClass("Rcpp_SHGInterface")
}
\keyword{classes}
