% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_SHGInterface}
\alias{Rcpp_SHGInterface}
\title{Rcpp SHG Interface Class}
\description{
This module provides an Rcpp interface to the Smoking History Generator (SHG) application, including intent-oriented config methods (\code{getConfig}/\code{useConfig}) and reproducibility export (\code{getReproConfig}).
}
\details{
Rcpp SHG Interface Class
}
