% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_SHGInterface}
\alias{Rcpp_SHGInterface}
\title{Test title SHGInterface}
\description{
TEST Type the name of the class to see its methods
}
\details{
Rcpp_SHGInterface Class
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor}
}}

