% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/shg-rcpp-docs.R
\name{Rcpp_SHGInterface}
\alias{Rcpp_SHGInterface}
\title{Rcpp SHG Interface Class}
\value{
The exported reference class \code{SHGInterface} (see \code{\link{SHGInterface}}):
an external pointer to the C++ engine used for all simulations.
}
\description{
This module provides an Rcpp interface to the Smoking History Generator (SHG) application, including intent-oriented config methods (\code{getConfig}/\code{useConfig}) and reproducibility export (\code{getReproConfig}).
}
\details{
Rcpp SHG Interface Class
}
