\name{RcppSmokingHistoryGenerator-package}
\alias{RcppSmokingHistoryGenerator-package}
\alias{RcppSmokingHistoryGenerator}
\docType{package}
\title{
\packageTitle{RcppSmokingHistoryGenerator}
}
\description{
\packageDescription{RcppSmokingHistoryGenerator}
}
\details{
Details sample
}
\author{
\packageAuthor{RcppSmokingHistoryGenerator}

Maintainer: \packageMaintainer{RcppSmokingHistoryGenerator}
}
\references{
    \enumerate{
        \item \href{https://resources.cisnet.cancer.gov/projects/#shg/tcpd}{Tobacco Control and the Reduction in Smoking-Related Premature Deaths in the United States, 1964-2012}
        \item \href{https://resources.cisnet.cancer.gov/projects/#shg/tce}{Patterns of Birth Cohort-Specific Smoking Histories, 1965-2009}
        \item \href{https://resources.cisnet.cancer.gov/projects/#shg/iomr}{Public Health Implications of Raising the Minimum Age of Legal Access to Tobacco Products}
        \item \href{https://resources.cisnet.cancer.gov/projects/#shg/sbc2}{Smoking and Lung Cancer Mortality in the United States From 2015 to 2065: A Comparative Modeling Approach}
    }
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = RcppSmokingHistoryGenerator)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{RcppSmokingHistoryGenerator}}
%\section{Documentation Index}{\packageIndices{RcppSmokingHistoryGenerator}}

\seealso{
CISNET Smoking History Generator source code: \url{https://github.com/CSNW/smoking-history-generator/}
}
