\name{RcppSmokingHistoryGenerator-package}
\alias{RcppSmokingHistoryGenerator-package}
\alias{RcppSmokingHistoryGenerator}
\docType{package}
\title{
\packageTitle{RcppSmokingHistoryGenerator}
}
\description{
\packageDescription{RcppSmokingHistoryGenerator}
}
\details{
Details sample
}
\author{
\packageAuthor{RcppSmokingHistoryGenerator}

Maintainer: \packageMaintainer{RcppSmokingHistoryGenerator}
}
\references{
References section
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = RcppSmokingHistoryGenerator)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{RcppSmokingHistoryGenerator}}
%\section{Documentation Index}{\packageIndices{RcppSmokingHistoryGenerator}}

\seealso{
See also section
}
