% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSimFromDataFrame}
\alias{runSimFromDataFrame}
\title{runSimFromDataFrame method}
\arguments{
\item{dfPopulation}{The input dataframe with named columns for race, sex, and birth_cohort}
}
\description{
runSimFromDataFrame offers a way to configure and run a simulation from an existing R dataframe. It returns a dataframe of simulated smoking histories with the same number of rows and order as the input dataframe.
}
\examples{
library(RcppSmokingHistoryGenerator)

shg <- new(SHGInterface)
N <- 10^6
pop <- list(
    race = rep(0, N),
    sex = sample(x = c(0, 1), size = N, prob = c(0.5, 0.5), replace = TRUE),
    birth_cohort = rep(1930:1949, N / 20)
)
shg$rng_strategy <- "RngStream"
shg$number_of_segments <- 10
shg$run_multi_threaded <- TRUE

smoking_history <- shg$runSimFromDataFrame(pop)
}
