% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runSimFromFixedValues}
\alias{runSimFromFixedValues}
\title{runSimFromFixedValues method}
\arguments{
\item{repeat}{The number of individuals to simulate}

\item{race}{(default = 0 and refers to all races combined)}

\item{sex}{(0 for male, 1, for female)}

\item{cohort_year}{(four digit birth cohort year)}
}
\description{
runSimFromFixedValues offers a way to configure and run a simulation from fixed values for race, sex, and birth year cohort rather than passing a data frame. It returns a dataframe of simulated smoking histories for n individuals.
}
\examples{
\dontrun{
library(SmokingHistoryGenerator)
shg <- new(SHGInterface)
shg$input_data_folder <- system.file("extdata", package="SmokingHistoryGenerator")
N <- 10^6
smoking_history <- shg$runSimFromFixedValues(N, 0, 0, 1950)
}
}
