\name{SmokingHistoryGenerator-package}
\alias{SmokingHistoryGenerator-package}
\alias{SmokingHistoryGenerator}
\docType{package}
\title{
\packageTitle{SmokingHistoryGenerator}
}
\description{
\packageDescription{SmokingHistoryGenerator}
}
\details{
Default calibrated inputs ship under \code{system.file("extdata", "2018", package = "SmokingHistoryGenerator")} as \code{smoking/*.csv} and \code{mort/*.csv} (NHIS-1965-2018 csv-partial). Set \code{input_data_folder} and per-table filenames accordingly. Wide legacy \code{.txt} tables remain supported. Full NHIS-style tables are distributed separately (Zenodo; the installed README is at \code{system.file("README.md", package = "SmokingHistoryGenerator")}).
}
\author{
\packageAuthor{SmokingHistoryGenerator}

Maintainer: \packageMaintainer{SmokingHistoryGenerator}
}
\references{
CISNET modelling applications that use the Smoking History Generator are listed
on the CISNET website at \url{https://cisnet.cancer.gov/}.
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SmokingHistoryGenerator)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SmokingHistoryGenerator}}
%\section{Documentation Index}{\packageIndices{SmokingHistoryGenerator}}

\seealso{
R package source: \url{https://github.com/NCI-CISNET/shg-r}
}
